requirements.txt
django == 3.2.8
sqlparse == 0.2.3
djangorestframework == 3.11.2
djangorestframework-jwt == 1.11.0
django-cors-headers == 3.10.0
MySQL-python == 1.2.5
pyyaml
